+++ packages/webrtc-core/src/rtc-peer-connection-factory.ts
@@ -3,10 +3,11 @@ import 'webrtc-adapter';
 /**
  * Creates an RTCPeerConnection.
  *
+ * @param configuration - Config to the RTCPeerConnection constructor.
  * @returns An RTCPeerConnection instance.
  */
-function createRTCPeerConnection(): RTCPeerConnection {
-  return new RTCPeerConnection();
+function createRTCPeerConnection(configuration?: RTCConfiguration | undefined): RTCPeerConnection {
+  return new RTCPeerConnection(configuration);
 }
 
 export { createRTCPeerConnection };