+++ src/media/local-microphone-track.ts
@@ -1,0 +1,6 @@
+import { LocalAudioTrack } from './local-audio-track';
+
+/**
+ * Represents a local track for a microphone source.
+ */
+export class LocalMicrophoneTrack extends LocalAudioTrack {}